chore: moves Multichain specs to tests
