+++ tests/helpers/multichain/connections/helpers.ts
@@ -1,5 +1,5 @@
-import { Matchers } from '../../../../tests/framework';
-import { createLogger } from '../../../../tests/framework/logger';
+import { Matchers } from '../../../framework';
+import { createLogger } from '../../../framework/logger';
 import { BrowserViewSelectorsIDs } from '../../../../app/components/Views/BrowserTab/BrowserView.testIds';
 
 const logger = createLogger({
