+++ tests/regression/multichain/permission-system-summary-default-permissions.spec.ts
@@ -1,12 +1,12 @@
-import { RegressionNetworkAbstractions } from '../../tags';
-import Browser from '../../pages/Browser/BrowserView';
-import ConnectedAccountsModal from '../../pages/Browser/ConnectedAccountsModal';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import { loginToApp, navigateToBrowserView } from '../../viewHelper';
-import Assertions from '../../../tests/framework/Assertions';
-import PermissionSummaryBottomSheet from '../../pages/Browser/PermissionSummaryBottomSheet';
-import { DappVariants } from '../../../tests/framework/Constants';
+import { RegressionNetworkAbstractions } from '../../../e2e/tags';
+import Browser from '../../../e2e/pages/Browser/BrowserView';
+import ConnectedAccountsModal from '../../../e2e/pages/Browser/ConnectedAccountsModal';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper';
+import { loginToApp, navigateToBrowserView } from '../../../e2e/viewHelper';
+import Assertions from '../../framework/Assertions';
+import PermissionSummaryBottomSheet from '../../../e2e/pages/Browser/PermissionSummaryBottomSheet';
+import { DappVariants } from '../../framework/Constants';
 import { PermissionSummaryBottomSheetSelectorsText } from '../../../app/components/Views/AccountConnect/PermissionSummaryBottomSheet.testIds';
 
 describe(
