+++ tests/regression/multichain/permissions/chains/permission-system-revoke-single.spec.js
@@ -1,13 +1,16 @@
-import TestHelpers from '../../../../helpers';
-import { RegressionNetworkExpansion } from '../../../../tags';
-import Browser from '../../../../pages/Browser/BrowserView';
-import NetworkListModal from '../../../../pages/Network/NetworkListModal';
-import ConnectedAccountsModal from '../../../../pages/Browser/ConnectedAccountsModal';
-import FixtureBuilder from '../../../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../../../tests/framework/fixtures/FixtureHelper';
-import { loginToApp, navigateToBrowserView } from '../../../../viewHelper';
-import Assertions from '../../../../../tests/framework/Assertions';
-import { DappVariants } from '../../../../../tests/framework/Constants';
+import TestHelpers from '../../../../../e2e/helpers';
+import { RegressionNetworkExpansion } from '../../../../../e2e/tags';
+import Browser from '../../../../../e2e/pages/Browser/BrowserView';
+import NetworkListModal from '../../../../../e2e/pages/Network/NetworkListModal';
+import ConnectedAccountsModal from '../../../../../e2e/pages/Browser/ConnectedAccountsModal';
+import FixtureBuilder from '../../../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../../../framework/fixtures/FixtureHelper';
+import {
+  loginToApp,
+  navigateToBrowserView,
+} from '../../../../../e2e/viewHelper';
+import Assertions from '../../../../framework/Assertions';
+import { DappVariants } from '../../../../framework/Constants';
 
 describe(RegressionNetworkExpansion('Chain Permission Management'), () => {
   beforeAll(async () => {
