+++ tests/smoke/multichain/permissions/chains/permission-system-dapp-chain-switch-grant.spec.js
@@ -1,18 +1,21 @@
-import FixtureBuilder from '../../../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../../../tests/framework/fixtures/FixtureHelper';
-import Browser from '../../../../pages/Browser/BrowserView';
-import ConnectBottomSheet from '../../../../pages/Browser/ConnectBottomSheet';
-import TestDApp from '../../../../pages/Browser/TestDApp';
-import { CustomNetworks } from '../../../../../tests/resources/networks.e2e';
-import { SmokeNetworkAbstractions } from '../../../../tags';
-import Assertions from '../../../../../tests/framework/Assertions';
-import { loginToApp, navigateToBrowserView } from '../../../../viewHelper';
-import ConnectedAccountsModal from '../../../../pages/Browser/ConnectedAccountsModal';
-import NetworkConnectMultiSelector from '../../../../pages/Browser/NetworkConnectMultiSelector';
-import NetworkNonPemittedBottomSheet from '../../../../pages/Network/NetworkNonPemittedBottomSheet';
-import { DappVariants } from '../../../../../tests/framework/Constants';
-import { setupRemoteFeatureFlagsMock } from '../../../../../tests/api-mocking/helpers/remoteFeatureFlagsHelper';
-import { remoteFeatureMultichainAccountsAccountDetailsV2 } from '../../../../../tests/api-mocking/mock-responses/feature-flags-mocks';
+import FixtureBuilder from '../../../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../../../framework/fixtures/FixtureHelper';
+import Browser from '../../../../../e2e/pages/Browser/BrowserView';
+import ConnectBottomSheet from '../../../../../e2e/pages/Browser/ConnectBottomSheet';
+import TestDApp from '../../../../../e2e/pages/Browser/TestDApp';
+import { CustomNetworks } from '../../../../resources/networks.e2e';
+import { SmokeNetworkAbstractions } from '../../../../../e2e/tags';
+import Assertions from '../../../../framework/Assertions';
+import {
+  loginToApp,
+  navigateToBrowserView,
+} from '../../../../../e2e/viewHelper';
+import ConnectedAccountsModal from '../../../../../e2e/pages/Browser/ConnectedAccountsModal';
+import NetworkConnectMultiSelector from '../../../../../e2e/pages/Browser/NetworkConnectMultiSelector';
+import NetworkNonPemittedBottomSheet from '../../../../../e2e/pages/Network/NetworkNonPemittedBottomSheet';
+import { DappVariants } from '../../../../framework/Constants';
+import { setupRemoteFeatureFlagsMock } from '../../../../api-mocking/helpers/remoteFeatureFlagsHelper';
+import { remoteFeatureMultichainAccountsAccountDetailsV2 } from '../../../../api-mocking/mock-responses/feature-flags-mocks';
 
 describe.skip(SmokeNetworkAbstractions('Chain Permission System'), () => {
   beforeAll(async () => {
