+++ tests/smoke/multichain/solana-wallet-standard/transferSol.spec.ts
@@ -1,11 +1,14 @@
-import { SmokeNetworkExpansion } from '../../../tags';
-import SolanaTestDApp from '../../../pages/Browser/SolanaTestDApp';
-import { connectSolanaTestDapp, navigateToSolanaTestDApp } from './testHelpers';
-import { loginToApp } from '../../../viewHelper';
-import { DappVariants } from '../../../../tests/framework/Constants';
-import FixtureBuilder from '../../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../../tests/framework/fixtures/FixtureHelper';
-import Assertions from '../../../../tests/framework/Assertions';
+import { SmokeNetworkExpansion } from '../../../../e2e/tags';
+import SolanaTestDApp from '../../../../e2e/pages/Browser/SolanaTestDApp';
+import {
+  connectSolanaTestDapp,
+  navigateToSolanaTestDApp,
+} from '../../../flows/solana-connection.flow';
+import { loginToApp } from '../../../../e2e/viewHelper';
+import { DappVariants } from '../../../framework/Constants';
+import FixtureBuilder from '../../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../../framework/fixtures/FixtureHelper';
+import Assertions from '../../../framework/Assertions';
 
 describe(
   SmokeNetworkExpansion('Solana Wallet Standard E2E - Transfer SOL'),
