+++ tests/smoke/multichain/wallet-revokeSession.spec.ts
@@ -2,12 +2,12 @@
  * E2E tests for wallet_revokeSession API
  * Tests revoking sessions and verifying the session state afterwards
  */
-import { SmokeMultiChainAPI } from '../../tags';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import MultichainTestDApp from '../../pages/Browser/MultichainTestDApp';
-import MultichainUtilities from '../../utils/MultichainUtilities';
-import { DappVariants } from '../../../tests/framework/Constants';
+import { SmokeMultiChainAPI } from '../../../e2e/tags';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper';
+import MultichainTestDApp from '../../../e2e/pages/Browser/MultichainTestDApp';
+import MultichainUtilities from '../../../e2e/utils/MultichainUtilities';
+import { DappVariants } from '../../framework/Constants';
 
 describe(SmokeMultiChainAPI('wallet_revokeSession'), () => {
   beforeEach(() => {
