+++ tests/smoke/multichain/wallet-sessionChanged.spec.ts
@@ -2,13 +2,13 @@
  * E2E tests for wallet_sessionChanged API
  * Tests that sessionChanged event is fired when networks are added to the session
  */
-import { SmokeMultiChainAPI } from '../../tags';
-import FixtureBuilder from '../../../tests/framework/fixtures/FixtureBuilder';
-import { withFixtures } from '../../../tests/framework/fixtures/FixtureHelper';
-import MultichainTestDApp from '../../pages/Browser/MultichainTestDApp';
-import MultichainUtilities from '../../utils/MultichainUtilities';
-import Assertions from '../../../tests/framework/Assertions';
-import { DappVariants } from '../../../tests/framework/Constants';
+import { SmokeMultiChainAPI } from '../../../e2e/tags';
+import FixtureBuilder from '../../framework/fixtures/FixtureBuilder';
+import { withFixtures } from '../../framework/fixtures/FixtureHelper';
+import MultichainTestDApp from '../../../e2e/pages/Browser/MultichainTestDApp';
+import MultichainUtilities from '../../../e2e/utils/MultichainUtilities';
+import Assertions from '../../framework/Assertions';
+import { DappVariants } from '../../framework/Constants';
 
 describe(SmokeMultiChainAPI('wallet_sessionChanged'), () => {
   beforeEach(() => {
